requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
dnspython==2.6.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.1
fiona==1.9.6
fonttools==4.51.0
geopandas==0.14.3
h11==0.14.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
pymongo==4.6.3
pyparsing==3.1.2
pyproj==3.6.1
pyspark==3.5.1
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==25.1.2
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
tornado==6.4
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
uvicorn==0.29.0
wcwidth==0.2.13
zipp==3.18.1
